refactor(assets): organize src/assets into logos/ + iconography/
Settle the asset directory convention (the source of the earlier
/src/assets vs /static confusion):
- src/assets/ = BUNDLED (imported in TSX, webpack-hashed, no stable URL)
  - logos/       per-stylesheet wordmarks (moved from flat src/assets/)
  - iconography/ JSX-rendered icons (scaffolded for incoming bookmark,
                 signal-on/off, letter-A, …)
- src/static/ = SERVED verbatim at /static/ (referenced by URL string)
- src/stylesheets/{name}/images/ = per-stylesheet served assets

Update logo imports in PrivateHeader + PublicLayout. Remove two unreferenced
strays (nina.png, default.jpg) left from the avatar-path experiment — the
real avatars live in src/static/common/avatars/. Add src/assets/README.md
documenting the import-vs-served rule.
