+++ src/assets/iconography/.gitkeep
@@ -1,0 +1,2 @@
+# Drop JSX-rendered icons here (bookmark.png, signal-on.png, signal-off.png,
+# letter-A.png, …). Imported in components; see ../README.md.
