+++ src/assets/logos/dark-ambient-logo.png
