+++ src/assets/logos/kuro-logo-hover.png
