+++ src/assets/logos/kuro-logo.png
