+++ src/assets/logos/layer-cake-logo.png
