+++ src/components/pages/private/layout/PrivateHeader.tsx
@@ -1,13 +1,13 @@
 import { useState } from 'react';
 import { Link, NavLink } from 'react-router-dom';
 import UserMenu from '../../../layout/UserMenu';
-import kuroLogo from '../../../../assets/kuro-logo.png';
-import kuroLogoHover from '../../../../assets/kuro-logo-hover.png';
-import darkAmbientLogo from '../../../../assets/dark-ambient-logo.png';
-import layerCakeLogo from '../../../../assets/layer-cake-logo.png';
-import layerCakeLogoHover from '../../../../assets/layer-cake-logo-hover.png';
-import postmodLogo from '../../../../assets/postmod-logo.png';
-import protonLogo from '../../../../assets/proton-logo.png';
+import kuroLogo from '../../../../assets/logos/kuro-logo.png';
+import kuroLogoHover from '../../../../assets/logos/kuro-logo-hover.png';
+import darkAmbientLogo from '../../../../assets/logos/dark-ambient-logo.png';
+import layerCakeLogo from '../../../../assets/logos/layer-cake-logo.png';
+import layerCakeLogoHover from '../../../../assets/logos/layer-cake-logo-hover.png';
+import postmodLogo from '../../../../assets/logos/postmod-logo.png';
+import protonLogo from '../../../../assets/logos/proton-logo.png';
 import Alert from '../../../layout/Alert';
 import ModBar from '../../../admin/ModBar';
 import QuickSearch from '../../../layout/QuickSearch';
@@ -24,7 +24,7 @@ import {
 } from '../../../../store/services/staffInboxApi';
 import { useGetMyProfileQuery } from '../../../../store/services/profileApi';
 
-// Asset naming convention: src/assets/{theme}-logo.png, src/assets/{theme}-logo-hover.png
+// Asset naming convention: src/assets/logos/{stylesheet}-logo.png (+ -hover.png)
 const THEME_LOGOS: Record<string, [string, string]> = {
   kuro: [kuroLogo, kuroLogoHover],
   'dark-ambient': [darkAmbientLogo, darkAmbientLogo],
